examples: convert to typescript.
